requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
PySide6==6.6.1
pyinstaller==6.3.0

